refactor(core): batch 7 final — Integer(int64_t(literal)) → Integer(literal)
7 more spots cleaned across:
- number_field.hpp: f_d_inv
- hensel_sqrt.hpp: f_lead_inv
- class_group.hpp: content (default ctor = 0)
- ecm.hpp: Point() = (Integer{}, 1), accum (×2)
- siqs.hpp: value, Q (default ctor = 0)

All Integer(int64_t(0)) replaced by default ctor, all (int64_t(1/2))
replaced by direct literal which dispatches to Integer(int) overload.

Smoke 28/28 PASS.

+++ include/gnfs/sqrt/hensel_sqrt.hpp
@@ -1346,7 +1346,7 @@ private:
             uint32_t d,
             const Integer& modulus) {
         // v22: f_d 直接 assign
-        Integer f_lead_inv(int64_t(1));
+        Integer f_lead_inv(1);
         Integer f_d;
         f_d = f[d];
         f_d %= modulus;
